create add and edit modal skeletons
